requirements.txt
tensorflow>=2.10
numpy
matplotlib
pillow
streamlit
scikit-learn
opencv-python
tqdm
pyyaml
scipy
